requirements.txt
boto3==1.4.1
botocore==1.4.93
Django==1.8
django-extensions==1.8.0
docutils==0.13.1
GDAL==2.1.3
jmespath==0.9.3
psycopg2==2.6.1
python-dateutil==2.6.0
python-slugify==1.2.4
s3transfer==0.1.10
six==1.10.0
Unidecode==0.4.21